moved ProjectNavigation into templates folder
