+++ src/components/index.js
@@ -1,8 +1,7 @@
-export { default as Button } from './button'
-export { default as Section } from './section'
-export { default as Card } from './card'
-export { default as Banner } from './banner'
-export { default as Header } from './header'
-export { default as Dropdown } from './dropdown'
-export { default as PageHeading } from './page-heading'
-export { default as ProjectNavigation } from './project-navigation'
+export { default as Button } from "./button"
+export { default as Section } from "./section"
+export { default as Card } from "./card"
+export { default as Banner } from "./banner"
+export { default as Header } from "./header"
+export { default as Dropdown } from "./dropdown"
+export { default as PageHeading } from "./page-heading"
