--- src/components/project-navigation/index.js
@@ -1,1 +1,0 @@
-export { default } from './ProjectNavigation'
