+++ src/templates/project-page.js
@@ -2,7 +2,8 @@ import React from "react"
 import { graphql } from "gatsby"
 import useAppStore from "../stores/store"
 import Layout from "../components/Layout"
-import { PageHeading, ProjectNavigation } from "../components"
+import { PageHeading } from "../components"
+import ProjectNavigation from "./ProjectNavigation"
 import { Box, Modal, Button, Typography } from "@mui/material"
 import styled from "styled-components"
